LPS-161984 Extract hook useObjectFieldForm from ObjectFieldFormBase component
