+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectField/EditObjectField.tsx
@@ -34,8 +34,9 @@ import {
 } from '../ModalAddFilter';
 import {AttachmentProperties} from './AttachmentProperties';
 import {MaxLengthProperties} from './MaxLengthProperties';
-import ObjectFieldFormBase, {useObjectFieldForm} from './ObjectFieldFormBase';
+import ObjectFieldFormBase from './ObjectFieldFormBase';
 import {SearchableContainer} from './SearchableContainer';
+import {useObjectFieldForm} from './useObjectFieldForm';
 
 import './EditObjectField.scss';
