+++ modules/apps/object/object-web/src/main/resources/META-INF/resources/js/components/ObjectField/ObjectFieldFormBase.tsx
@@ -21,10 +21,7 @@ import {
 	Input,
 	Select,
 	SingleSelect,
-	invalidateRequired,
-	useForm,
 } from '@liferay/object-js-components-web';
-import {sub} from 'frontend-js-web';
 import React, {
 	ChangeEventHandler,
 	ReactNode,
@@ -33,7 +30,6 @@ import React, {
 	useState,
 } from 'react';
 
-import {normalizeFieldSettings} from '../../utils/fieldSettings';
 import {toCamelCase} from '../../utils/string';
 import {AggregationFormBase} from './AggregationFormBase';
 import {AttachmentFormBase} from './AttachmentFormBase';
@@ -41,7 +37,6 @@ import {AttachmentFormBase} from './AttachmentFormBase';
 import './ObjectFieldFormBase.scss';
 
 const defaultLanguageId = Liferay.ThemeDisplay.getDefaultLanguageId();
-const REQUIRED_MSG = Liferay.Language.get('required');
 
 export default function ObjectFieldFormBase({
 	children,
@@ -374,185 +369,6 @@ export default function ObjectFieldFormBase({
 		</>
 	);
 }
-
-export function useObjectFieldForm({
-	forbiddenChars,
-	forbiddenLastChars,
-	forbiddenNames,
-	initialValues,
-	onSubmit,
-}: IUseObjectFieldForm) {
-	const validate = (field: Partial<ObjectField>) => {
-		const getSourceFolderError = (folderPath: string) => {
-
-			// folder name cannot end with invalid last characters
-
-			const lastChar = folderPath[folderPath.length - 1];
-
-			if (forbiddenLastChars?.some((char) => char === lastChar)) {
-				return sub(
-					Liferay.Language.get(
-						'the-folder-name-cannot-end-with-the-following-characters-x'
-					),
-					forbiddenLastChars.join(' ')
-				);
-			}
-
-			// folder name cannot contain invalid characters
-
-			if (forbiddenChars?.some((symbol) => folderPath.includes(symbol))) {
-				return sub(
-					Liferay.Language.get(
-						'the-folder-name-cannot-contain-the-following-invalid-characters-x'
-					),
-					forbiddenChars.join(' ')
-				);
-			}
-
-			// folder name cannot be a reserved word
-
-			const reservedNames = new Set(forbiddenNames);
-
-			if (
-				forbiddenNames &&
-				folderPath.split('/').some((name) => reservedNames.has(name))
-			) {
-				return sub(
-					Liferay.Language.get(
-						'the-folder-name-cannot-have-a-reserved-word-such-as-x'
-					),
-					forbiddenNames.join(', ')
-				);
-			}
-
-			return null;
-		};
-
-		const errors: ObjectFieldErrors = {};
-
-		const label = field.label?.[defaultLanguageId];
-
-		const settings = normalizeFieldSettings(field.objectFieldSettings);
-
-		if (invalidateRequired(label)) {
-			errors.label = REQUIRED_MSG;
-		}
-
-		if (invalidateRequired(field.name ?? label)) {
-			errors.name = REQUIRED_MSG;
-		}
-
-		if (!field.businessType) {
-			errors.businessType = REQUIRED_MSG;
-		}
-		else if (field.businessType === 'Aggregation') {
-			if (!settings.function) {
-				errors.function = REQUIRED_MSG;
-			}
-
-			if (settings.function !== 'COUNT' && !settings.objectFieldName) {
-				errors.objectFieldName = REQUIRED_MSG;
-			}
-
-			if (!settings.objectRelationshipName) {
-				errors.objectRelationshipName = REQUIRED_MSG;
-			}
-		}
-		else if (field.businessType === 'Attachment') {
-			const uploadRequestSizeLimit = Math.floor(
-				Liferay.PropsValues.UPLOAD_SERVLET_REQUEST_IMPL_MAX_SIZE /
-					1048576
-			);
-
-			if (
-				invalidateRequired(
-					settings.acceptedFileExtensions as string | undefined
-				)
-			) {
-				errors.acceptedFileExtensions = REQUIRED_MSG;
-			}
-			if (!settings.fileSource) {
-				errors.fileSource = REQUIRED_MSG;
-			}
-			if (!settings.maximumFileSize && settings.maximumFileSize !== 0) {
-				errors.maximumFileSize = REQUIRED_MSG;
-			}
-			else if (settings.maximumFileSize > uploadRequestSizeLimit) {
-				errors.maximumFileSize = sub(
-					Liferay.Language.get(
-						'file-size-is-larger-than-the-allowed-overall-maximum-upload-request-size-x-mb'
-					),
-					uploadRequestSizeLimit
-				);
-			}
-			else if (settings.maximumFileSize < 0) {
-				errors.maximumFileSize = sub(
-					Liferay.Language.get(
-						'only-integers-greater-than-or-equal-to-x-are-allowed'
-					),
-					0
-				);
-			}
-
-			if (settings.showFilesInDocumentsAndMedia) {
-				if (
-					invalidateRequired(
-						settings.storageDLFolderPath as string | undefined
-					)
-				) {
-					errors.storageDLFolderPath = REQUIRED_MSG;
-				}
-				else {
-					const sourceFolderError = getSourceFolderError(
-						settings.storageDLFolderPath as string
-					);
-
-					if (sourceFolderError !== null) {
-						errors.storageDLFolderPath = sourceFolderError;
-					}
-				}
-			}
-		}
-		else if (
-			field.businessType === 'Text' ||
-			field.businessType === 'LongText'
-		) {
-			if (settings.showCounter && !settings.maxLength) {
-				errors.maxLength = REQUIRED_MSG;
-			}
-		}
-		else if (field.businessType === 'Picklist') {
-			if (!field.listTypeDefinitionId) {
-				errors.listTypeDefinitionId = REQUIRED_MSG;
-			}
-
-			if (field.state && !field.defaultValue) {
-				errors.defaultValue = REQUIRED_MSG;
-			}
-		}
-
-		return errors;
-	};
-
-	const {errors, handleChange, handleSubmit, setValues, values} = useForm<
-		ObjectField,
-		{[key in ObjectFieldSettingName]: unknown}
-	>({
-		initialValues,
-		onSubmit,
-		validate,
-	});
-
-	return {errors, handleChange, handleSubmit, setValues, values};
-}
-interface IUseObjectFieldForm {
-	forbiddenChars?: string[];
-	forbiddenLastChars?: string[];
-	forbiddenNames?: string[];
-	initialValues: Partial<ObjectField>;
-	onSubmit: (field: ObjectField) => void;
-}
-
 interface IProps {
 	children?: ReactNode;
 	disabled?: boolean;
